Add files via upload
The DDoS attack is detected by considering the various SNMP parameters and classifying using machine learning technique like bagging, boosting and ensemble models. Also, the various types of malware on the networked devices are prevent from being used as a bot for DDOS attack generation.Distributed Denial of Service (DDoS) attacks falls in the category of critical attacks that compromises the availability of the network resources and detection of these attacks is also a challenging task. The objective of this paper is to develop an Alert Classification System using Machine Learning Algorithms Artificial Neural Networks and Support Vector Machines. An Experimental testbed with 20 nodes and a Server is used to Generate, Detect and Classify these Attacks along with normal Traffic. The System is tested with real traffic and the classification Accuracy is found to be greater than the Rule based and Threshold Methods.